trait detail bug fix
trait detail didn't update correctly when toggling a trait on/off and
adding/removing a skill with that trait between toggles. fire an update
to trait detail when toggling for the same trait.

additionaly close trait detail when the build changes. a better solution
would be to udpate the trait detail, but this is quicker and easier for
now.